@@ -23,29 +23,31 @@
     "validate": "yarn type-check && yarn lint && yarn lint:css && yarn test"
   },
   "dependencies": {
-    "@mdx-js/loader": "1.0.4",
-    "@mdx-js/react": "1.0.2",
-    "@next/mdx": "8.0.4",
-    "@sentry/browser": "5.0.8",
-    "@sentry/node": "5.0.8",
+    "@mdx-js/loader": "1.0.19",
+    "@mdx-js/react": "1.0.16",
+    "@next/mdx": "8.1.0",
+    "@sentry/browser": "5.2.1",
+    "@sentry/node": "5.2.0",
+    "@types/intl-relativeformat": "2.1.0",
     "@zeit/next-bundle-analyzer": "0.1.2",
     "@zeit/next-source-maps": "0.0.3",
     "@zeit/next-typescript": "1.1.1",
     "body-scroll-lock": "2.6.1",
     "date-fns": "2.0.0-alpha.27",
+    "intl": "1.2.5",
+    "intl-relativeformat": "2.1.0",
     "is-absolute-url": "3.0.0",
-    "next": "8.0.5-canary.33",
+    "next": "8.1.1-canary.29",
     "next-ga": "2.3.4",
     "next-nprogress": "1.4.0",
-    "next-server": "8.0.5-canary.33",
+    "next-server": "8.1.1-canary.29",
     "nprogress": "0.2.0",
-    "polished": "3.2.0",
+    "polished": "3.3.0",
     "prism-react-renderer": "0.1.6",
     "react": "16.8.6",
     "react-dom": "16.8.6",
-    "react-intl": "2.8.0",
-    "react-simple-img": "2.1.8",
-    "react-spring": "8.0.19",
+    "react-simple-img": "2.1.9",
+    "react-spring": "8.0.20",
     "somebody": "2.0.0",
     "styled-components": "4.2.0"
   },
@@ -55,62 +57,59 @@
     "@types/common-tags": "1.8.0",
     "@types/glob": "7.1.1",
     "@types/is-absolute-url": "3.0.0",
-    "@types/jest": "24.0.11",
+    "@types/jest": "24.0.13",
     "@types/jest-in-case": "1.0.1",
-    "@types/next": "8.0.3",
     "@types/nprogress": "0.0.29",
-    "@types/prettier": "1.16.1",
-    "@types/react": "16.8.13",
+    "@types/prettier": "1.16.4",
+    "@types/react": "16.8.17",
     "@types/react-dom": "16.8.4",
-    "@types/react-intl": "2.3.17",
-    "@types/styled-components": "4.1.14",
-    "@typescript-eslint/eslint-plugin": "1.6.0",
-    "@typescript-eslint/parser": "1.6.0",
+    "@types/styled-components": "4.1.15",
+    "@typescript-eslint/eslint-plugin": "1.9.0",
+    "@typescript-eslint/parser": "1.9.0",
     "babel-core": "7.0.0-bridge.0",
-    "babel-jest": "24.7.1",
+    "babel-jest": "24.8.0",
     "babel-plugin-inline-react-svg": "1.1.0",
-    "babel-plugin-react-intl": "3.0.1",
-    "babel-plugin-root-import": "6.1.0",
+    "babel-plugin-root-import": "6.2.0",
     "babel-plugin-styled-components": "1.10.0",
     "common-tags": "1.8.0",
     "eslint": "5.16.0",
     "eslint-config-airbnb": "17.1.0",
-    "eslint-config-kentcdodds": "14.3.1",
+    "eslint-config-kentcdodds": "14.3.2",
     "eslint-config-mcansh": "1.1.1",
-    "eslint-config-prettier": "4.1.0",
+    "eslint-config-prettier": "4.2.0",
     "eslint-import-resolver-babel-plugin-root-import": "1.1.1",
     "eslint-import-resolver-typescript": "1.1.1",
     "eslint-plugin-compat": "3.1.1",
-    "eslint-plugin-import": "2.17.1",
-    "eslint-plugin-jest": "22.4.1",
+    "eslint-plugin-import": "2.17.2",
+    "eslint-plugin-jest": "22.5.1",
     "eslint-plugin-jsx-a11y": "6.2.1",
     "eslint-plugin-mdx": "1.0.1",
-    "eslint-plugin-prettier": "3.0.1",
+    "eslint-plugin-prettier": "3.1.0",
     "eslint-plugin-promise": "4.1.1",
-    "eslint-plugin-react": "7.12.4",
+    "eslint-plugin-react": "7.13.0",
     "eslint-plugin-react-hooks": "1.6.0",
-    "glob": "7.1.3",
-    "husky": "1.3.1",
-    "jest": "24.7.1",
-    "jest-dom": "3.1.3",
+    "glob": "7.1.4",
+    "husky": "2.3.0",
+    "jest": "24.8.0",
+    "jest-dom": "3.2.2",
     "jest-in-case": "1.0.2",
     "jest-runner-eslint": "0.7.3",
-    "jest-runner-tsc": "1.5.0",
+    "jest-runner-tsc": "1.5.1",
     "jest-styled-components": "6.3.1",
     "kleur": "3.0.3",
-    "lint-staged": "8.1.5",
-    "next-offline": "3.3.8",
-    "nodemon": "1.18.11",
-    "prettier": "1.17.0",
-    "react-testing-library": "6.1.2",
-    "stylelint": "9.10.1",
+    "lint-staged": "8.1.7",
+    "next-offline": "4.0.0",
+    "nodemon": "1.19.0",
+    "prettier": "1.17.1",
+    "react-testing-library": "7.0.0",
+    "stylelint": "10.0.1",
     "stylelint-config-amp": "0.5.0",
     "stylelint-config-recommended": "2.2.0",
     "stylelint-config-styled-components": "0.1.1",
-    "stylelint-processor-styled-components": "1.6.0",
-    "ts-node": "8.0.3",
-    "typescript": "3.4.3",
-    "webpack": "4.30.0"
+    "stylelint-processor-styled-components": "1.7.0",
+    "ts-node": "8.1.0",
+    "typescript": "3.4.5",
+    "webpack": "4.31.0"
   },
   "greenkeeper": {
     "commitMessages": {
