@@ -1,30 +1,19 @@
-// @ts-ignore
 import { meta as learningSoftwareDevelopment } from './pages/why-im-learning-software-development.mdx';
-// @ts-ignore
 import { meta as html5Progress } from './pages/html5-progress-element.mdx';
-// @ts-ignore
 import { meta as objectLifecycle } from './pages/object-lifecycle-cheatsheet.mdx';
-// @ts-ignore
-import { meta as javascriptlasses } from './pages/javascript-classes.mdx';
-// @ts-ignore
+import { meta as javascriptClasses } from './pages/javascript-classes.mdx';
 import { meta as rackKeyConcepts } from './pages/rack-key-concepts.mdx';
-// @ts-ignore
 import { meta as lifeAndCode } from './pages/similarities-between-life-and-code.mdx';
-// @ts-ignore
 import { meta as timeToHex } from './pages/time-to-hex.mdx';
-// @ts-ignore
 import { meta as sinatraProject } from './pages/sinatra-project.mdx';
-// @ts-ignore
 import { meta as railsProject } from './pages/rails-project.mdx';
-// @ts-ignore
 import { meta as rubyProject } from './pages/ruby-project.mdx';
-import { Post } from '~/components/post-card/index';
 
-const posts = [
+export default [
   learningSoftwareDevelopment,
   html5Progress,
   objectLifecycle,
-  javascriptlasses,
+  javascriptClasses,
   rackKeyConcepts,
   lifeAndCode,
   timeToHex,
@@ -32,15 +21,3 @@ const posts = [
   railsProject,
   rubyProject,
 ];
-
-const sortPosts = (postsArray: Post[]) =>
-  postsArray.sort((a, b) => {
-    if (a.date > b.date) return -1;
-    if (a.date < b.date) return 1;
-    return 0;
-  });
-
-const sortedByDate = sortPosts(posts);
-
-export { sortPosts, sortedByDate };
-export default posts;
