@@ -1,20 +1,14 @@
 /* eslint-env jest */
 import React from 'react';
-import { IntlProvider } from 'react-intl';
 import { ThemeProvider } from 'styled-components';
 import { render } from 'react-testing-library';
 import theme from '~/config';
 
-const renderWithIntl = (
-  children: React.ReactNode,
-  intlProps = { locale: 'en' }
-) =>
+const renderWithIntl = (children: React.ReactNode) =>
   render(
-    <IntlProvider {...intlProps}>
-      <ThemeProvider theme={theme}>
-        <>{children}</>
-      </ThemeProvider>
-    </IntlProvider>
+    <ThemeProvider theme={theme}>
+      <>{children}</>
+    </ThemeProvider>
   );
 
 export * from 'react-testing-library';
